re: #248516 - Created Triggers won't edit
Fixes problem with editing child triggers. If the event is a task result, then it would not load attributes at all (crashes with Javascript error). Now a call to triggers/taskResults is made when there is a parent trigger, and the setup of the trigger is not done until that is complete. The propertyPage is not needed *at all*, as the change even on the task forces this to rebuild in all circumstances I could find.