Another attempt at fixing at the problem where triggers that belonged to automations could not be edited at all, this time without messing up 'ordinary' triggers and fixing the other issues discovered while tracking this down.
Some points to note :-

 * Timing IS involved. The display of the properties depends on the value of the 'target event' and 'target trigger'. These are loaded asynchronously, so prior to this changes whether or not the properties loaded was luck.

 * Prior to this change, there were a number of situations where the properties were loaded twice. At best this is just inefficient, at worst you would see a flicker as they properties were reloaded again.

 * The loading of property values wasn't happening because the 'changed' event was not getting fired as expected. This patch continues along the same lines as using the changed event as the trigger for loading new property values. This keeps this code down to being only called in one place, and fixes the doubling-loading problem.

 * I found a situation where child triggers of automations could become detached (by losing the attachedId). This patch also addresses that.

I have tested this change in both Automations and Triggers. You can happily create and edit both types, including children of children.

There was one oddity I saw once, but could not reproduce. When trying to create a child of a child of an automation, it seemed to attach to the wrong parent. However, I have not since been able to reproduce this and at the time my data was in a bit of mess by this point.