TinyMCE editors were missing expected functionality.
Despite the richInput input have a 'plugins' in options, and a specific list of toolbar items, it did not pass the plugins down to TinyMCE itself.

In order not to change existing uses of TinyMCE (I don't know where they), I added a default option 'usePlugins' that defaults to false. I then set this to true where I need the full featured editor. If the intention was to always have the full featured editor in richInput this default could be changed.

I also added 'extraPlugins' that adds to the additional list, and 'tinyMCEOptions' which allows ANY tinyMCE option to be passed down if needed.