#LB24 - Stylesheet field appears blank until you click in it
This was a race condition related to the requirement of CodeMirror to have a call to refresh() when the editor is actually attached to the DOM. Because at the point of hypersocket-properties building up the HTML for the dynamic form, this is not the case for many situations (e.g. when the form is in tabs). There were a few hacks in various places that try to call this (e.g. when a tab switched), but these prove ineffective in this case and are terrible unreliable in all other cases. It also set a timer that after 1ms tried to do the refresh. This is just not sufficient. Some pages may take multiple seconds to have a fully populated DOM.

So instead, the CodeMirror autorefresh addon has been included. This *polls* for the existence of the editor. When it exists, it will refresh and stop polling. This appears to be an infinitely more reliable method.