In order to move the ROI 'chart' options to the Dashboard options, this area has had to be split up. The existing dashboard options area was not extendable in any way (could not use a generic properties category), and does not use the property system to store preferences, instead using the UserInterfaceStateService and doing things it's own way.
This would be considerably more work to change this to use the generic architecture. I considered attempting to merge property tabs with generic content instead, but again this seemed a bit silly.

So extensions can now register things for placement in the new page. The special 'Layout' tab is separate and has had tab decorations removed as they were never used.