M628 - Narrow screen on Users page, user list gets pushed off the bottom of the screen
Added overflow-x to auto in div with id 'content', this pushes the overflow down to the containers. There is a deep hierarchy of dom nodes below 'content', but with this the table overflow down the hierarchy is triggered, as table is also marked with overflow, we start to see scrolls in table and the content stops to shift down as described in the ticket.