M660 - Cannot remove a user from a role on the users page
In principalRoleMapping.html using event.currentTarget as when we click on the cross icon, the event handler is attached to parent span which holds the data with key 'value'. event.target points to cross actually we need the parent which is given by event.currentTarget where the event handler is attached not where we clicked.

In hypersocket-widgets.js, using new syntax for attaching click handler the earlier ones is deprecated.