#M743 - Upgrade bootstrap to latest (Widget Dropdown click handling issue)
The custom dop down we create needs special handling hence added new class lb-dropdown-menu as a marker class, we use this to add document click handler.

If not done it conflicts with new bootstrap 5 dropdown handling, which also attached document on click to close drop down when clicked outside. SO with marker we distinguish.

Changes for how input group items are handled in bootstrap 5.